Hookify useWillRedirectAfterSuccess
