+++ packages/launch/src/hooks/index.ts
@@ -7,3 +7,4 @@ export * from './use-title';
 export * from './use-site-domains';
 export * from './use-plans';
 export * from './use-cart';
+export * from './use-will-redirect-after-success';
